[INLONG-2274][Manager] Move BusinessServiceImplTest.java to BusinessServiceTest.java
